dev-requirements.txt
alembic==1.15.2
babel==2.17.0
Beaker==1.13.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckan==2.11.3
click==8.2.0
cryptography==45.0.2
decorator==5.2.1
distlib==0.3.8
dominate==2.9.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.3.0
filelock==3.14.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
FormEncode==2.1.1
greenlet==3.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
msgspec==0.19.0
nose==1.3.7
packaging==25.0
passlib==1.7.4
Paste==3.10.1
PasteDeploy==3.1.0
PasteScript==3.7.0
pipenv==2023.12.1
platformdirs==4.2.2
pluggy==1.5.0
polib==1.2.0
pycparser==2.22
Pygments==2.19.1
Pylons==1.0.3
pyparsing==3.2.3
pysolr==3.10.0
pytest==8.3.5
pytest-ckan==0.0.12
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
repoze.lru==0.7
requests==2.32.3
Routes==2.5.1
rq==1.10.1
setuptools==80.8.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
Tempita==0.6.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
virtualenv==20.26.2
waitress==3.0.2
webassets==2.0
webencodings==0.5.1
WebError==0.13.1
WebHelpers==1.3
WebOb==1.8.9
WebTest==3.0.4
Werkzeug==3.1.3
wheel==0.43.0
WTForms==3.2.1
